fix subclass inheritance of elk attributes
Make the metaclass must look for elk attributes in the base classes
as well as in the class dict in order that subclasses will properly
inherit elk attributes.  Bases are traversed in reverse order before
the dict is checked.

Once the attribute list is complete, preserve it in the dict and
reuse it (non-destructively) during instantiation.

Add a test for elk attribute inheritance.