fix inheritance of elk roles
The attributes and "does" relationship are not correctly inherited
by descendents of ElkRole consumers.  Add some tests and fix the bug
by implementing ElkRole.__subclasscheck__ and delegating to
issubclass in the ``does`` method, which is now a classmethod.

Also add ElkRole.__instancecheck__ for consistency and add
assertions for isinstance and issubclass to the appropriate tests.