add spec for modifier-role combinations
Part of the usefulness of method modifiers is that should be able to
modify methods provided by a role, and that a role should be able to
modify methods of a class into which it is composed.  These
scenarios were not tested, so add some basic specs (happily, they
pass without changes).